Remove drive/control module and rewire references
Delete drive/control submodule files and update imports/usages.
Repointed drive::control::RotationState references to
drive::rotation::RotationState,
adjusted docs and types comments, and updated a test to use
DriveAction::Coast.
