--- src/task/drive/control/mod.rs
@@ -1,17 +1,0 @@
-//! Control systems for drive task
-//!
-//! This module contains all the control algorithms and state machines
-//! used for precise motion control, including rotation, straight-line
-//! driving, tilt compensation, and motor state management.
-//!
-//! Completion handles are resolved by the drive task (`drive/mod.rs`) after
-//! these control routines report progress or terminal states. The control
-//! modules do not own completion senders.
-
-pub mod motor_state;
-pub mod rotation;
-pub mod straight_line;
-pub mod tilt;
-
-// Re-export commonly used types for convenience
-pub use rotation::RotationState;
